requirements.txt
requests
redis
pymysql
DBUtils
binance-connector